Move lib code related to data collection to shared dir
