+++ src/lib/assets/destroy.ts
@@ -3,7 +3,7 @@
 
 import { K8s, kind } from "kubernetes-fluent-client";
 
-import Log from "../logger";
+import Log from "../telemetry/logger";
 import { peprStoreCRD } from "./store";
 
 export async function destroyModule(name: string) {
