+++ src/lib/telemetry/logger.test.ts
