+++ src/lib/telemetry/metrics.test.ts
